+++ sharding-transaction/sharding-transaction-xa/src/main/java/io/shardingsphere/transaction/xa/wrap/extractor/DataSourceParameterExtractorAdapter.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.xa.manager.extractor;
+package io.shardingsphere.transaction.xa.wrap.extractor;
 
 import io.shardingsphere.core.config.DataSourceConfiguration;
 import io.shardingsphere.core.rule.DataSourceParameter;
