+++ sharding-transaction/sharding-transaction-xa/src/test/java/io/shardingsphere/transaction/xa/manager/AllManagerTests.java
@@ -18,8 +18,9 @@
 package io.shardingsphere.transaction.xa.manager;
 
 import io.shardingsphere.transaction.xa.manager.atomikos.AtomikosTransactionManagerTest;
-import io.shardingsphere.transaction.xa.manager.extractor.DataSourceParameterFactoryTest;
-import io.shardingsphere.transaction.xa.manager.dialect.XAPropertyFactoryTest;
+import io.shardingsphere.transaction.xa.wrap.extractor.DataSourceParameterFactoryTest;
+import io.shardingsphere.transaction.xa.wrap.dialect.XAPropertyFactoryTest;
+import io.shardingsphere.transaction.xa.wrap.XATransactionDataSourceWrapperTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
