+++ sharding-transaction/sharding-transaction-xa/src/test/java/io/shardingsphere/transaction/xa/wrap/XATransactionDataSourceWrapperTest.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.xa.manager;
+package io.shardingsphere.transaction.xa.wrap;
 
 import com.atomikos.beans.PropertyException;
 import com.atomikos.jdbc.AtomikosDataSourceBean;
@@ -23,10 +23,9 @@ import io.shardingsphere.core.constant.DatabaseType;
 import io.shardingsphere.core.constant.transaction.ProxyPoolType;
 import io.shardingsphere.core.rule.DataSourceParameter;
 import io.shardingsphere.transaction.xa.XATransactionManagerSPILoader;
-import io.shardingsphere.transaction.xa.manager.atomikos.XATransactionDataSourceWrapper;
-import io.shardingsphere.transaction.xa.manager.dialect.XADataSourceFactory;
+import io.shardingsphere.transaction.xa.wrap.dialect.XADataSourceFactory;
 import io.shardingsphere.transaction.xa.manager.fixture.ReflectiveUtil;
-import io.shardingsphere.transaction.xa.manager.dialect.XADatabaseType;
+import io.shardingsphere.transaction.xa.wrap.dialect.XADatabaseType;
 import org.apache.tomcat.dbcp.dbcp2.managed.BasicManagedDataSource;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
