+++ sharding-transaction/sharding-transaction-xa/src/test/java/io/shardingsphere/transaction/xa/wrap/dialect/XAPropertyFactoryTest.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.xa.manager.dialect;
+package io.shardingsphere.transaction.xa.wrap.dialect;
 
 import io.shardingsphere.core.rule.DataSourceParameter;
 import org.junit.Before;
